+++ crates/node-resource-metrics/src/collectors/mod.rs
@@ -11,10 +11,15 @@ mod process_metrics_collector;
 
 pub(crate) use common::CollectorLatencyCollector;
 pub(crate) use cpu_metrics_collector::CpuMetricsCollector;
-pub(crate) use cpu_metrics_collector::LinuxCpuMetricsCollector;
 pub(crate) use disk_metrics_collector::DiskMetricsCollector;
-pub(crate) use disk_metrics_collector::LinuxDiskMetricsCollector;
 pub(crate) use loadavg_collector::LoadAvgCollector;
 pub(crate) use memory_metrics_collector::MemoryMetricsCollector;
 pub(crate) use network_metrics_collector::NetworkMetricsCollector;
 pub(crate) use process_metrics_collector::ProcessMetricsCollector;
+
+#[cfg(all(target_os = "linux"))]
+mod linux_collectors;
+#[cfg(all(target_os = "linux"))]
+pub(crate) use linux_collectors::LinuxCpuMetricsCollector;
+#[cfg(all(target_os = "linux"))]
+pub(crate) use linux_collectors::LinuxDiskMetricsCollector;
