+++ crates/node-resource-metrics/src/lib.rs
@@ -3,11 +3,9 @@
 
 use cfg_if::cfg_if;
 use collectors::{
-    CpuMetricsCollector, DiskMetricsCollector, LinuxCpuMetricsCollector, LinuxDiskMetricsCollector,
-    LoadAvgCollector, MemoryMetricsCollector, NetworkMetricsCollector, ProcessMetricsCollector,
+    CollectorLatencyCollector, CpuMetricsCollector, DiskMetricsCollector, LoadAvgCollector,
+    MemoryMetricsCollector, NetworkMetricsCollector, ProcessMetricsCollector,
 };
-
-use crate::collectors::CollectorLatencyCollector;
 
 mod collectors;
 
@@ -21,8 +19,8 @@ pub fn register_node_metrics_collector() {
     prometheus::register(Box::new(ProcessMetricsCollector::default())).unwrap();
     cfg_if! {
         if #[cfg(all(target_os="linux"))] {
-            prometheus::register(Box::new(LinuxCpuMetricsCollector::default())).unwrap();
-            prometheus::register(Box::new(LinuxDiskMetricsCollector::default())).unwrap();
+            prometheus::register(Box::new(collectors::LinuxCpuMetricsCollector::default())).unwrap();
+            prometheus::register(Box::new(collectors::LinuxDiskMetricsCollector::default())).unwrap();
         }
     }
     prometheus::register(Box::new(CollectorLatencyCollector::default())).unwrap();
